Restart server if it double-forks
Processes that double-fork themselves by double-forking cannot be waited
upon, and cannot be properly monitored.

The ARK server will sometimes restart itself by performing a fork and
execve instead of just an execve (presumably because that is what it has
to do under Windows).

If the server restarts itself, then we need to restart it again so it
can be properly monitored.